requirements.txt
pandas
numpy
streamlit
opencv-python
opencv-contrib-python
pandas-stubs
types-Pillow
pillow-avif-plugin
xlsxwriter
matplotlib